Implemented essential auth mechanism
Implemented the essential part of the auth mechanism: when connected, the server will send an AUTH request with the right parameter (Password is not implemented for now). Client will then ask the user to write a username and it will send it to the server. If the username does not already exist on the server, the user will be registered and a OK command will be sent. Otherwise, an AUTHFAIL command will be sent.